popular tags features added

+++ frontend/screens/SearchScreen.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { SafeAreaView, View, Text, StyleSheet, ScrollView, FlatList } from 'react-native';
+import { SafeAreaView, View, Text, StyleSheet, ScrollView, FlatList, Alert } from 'react-native';
 import { SearchBar, ListItem, Icon } from 'react-native-elements';
 import { useDispatch, useSelector } from 'react-redux';
 
@@ -30,15 +30,24 @@ const SearchScreen = props => {
 
     const deleteNote = async (noteTobeDeleted) => {
 
-        try {
-            await dispatch(notesActions.deleteNote(noteTobeDeleted))
-
-
-        } catch (err) {
-            console.log(err.message)
-            setError(err.message)
-
-        }
+        Alert.alert('Are you sure?', 'Do you really want to delete this item?', [
+            { text: 'No', style: 'default' },
+            {
+              text: 'Yes',
+              style: 'destructive',
+              onPress: async () => {
+                try {
+                    await dispatch(notesActions.deleteNote(noteTobeDeleted))
+        
+        
+                } catch (err) {
+                    console.log(err.message)
+                    setError(err.message)
+        
+                }
+              }
+            }
+          ]);
     }
 
     const renderHiddenItem = (itemData, itemMap) => (
